Update aws-properties-s3-bucket-defaultretention.md
The note has to be added, for visibility. It is the same note we have on the ObjectLockConfiguration.
If none of the time properties is added, as they are all marked Required: No, S3 returns XML conflict error

Note
DefaultRetention requires either Days or Years. You can't specify both at the same time.